# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "YANFEI API",
        "contact": {},
        "license": {},
        "version": "0.0.1"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/wx/work/publish": {
            "post": {
                "description": "发布工作",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作相关"
                ],
                "summary": "发布工作",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "工种 0(点工),1(包工),2(突击队) 必填",
                        "name": "work_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "点工招聘",
                        "name": "点工示例数据",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.DianWorkWrapper"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controller.Message"
                        }
                    }
                }
            }
        },
        "/wx/work/search": {
            "get": {
                "description": "搜索工作，需要某个筛选参数就加上，否则可以不加；按发布时间降序排序",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作相关"
                ],
                "summary": "搜索工作",
                "parameters": [
                    {
                        "type": "string",
                        "description": "二级位置信息 选填",
                        "name": "location",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "所需工种 选填",
                        "name": "need",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "工程类别 选填",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "工作类别 选填0只返回点工和包工，1只返回突击队，默认为0",
                        "name": "work_type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码，从1开始 必填",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页记录数 必填",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controller.Message"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.Message": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "model.DianWorkWrapper": {
            "type": "object",
            "properties": {
                "construction_company": {
                    "type": "string",
                    "example": "飞燕工程队"
                },
                "desc": {
                    "type": "string",
                    "example": "包吃包住"
                },
                "final_treatment": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "location": {
                    "type": "string",
                    "example": "湖北省襄阳市"
                },
                "location_info": {
                    "type": "object",
                    "$ref": "#/definitions/model.LocationInfoWrapper"
                },
                "max_wage": {
                    "type": "string",
                    "example": "200"
                },
                "min_wage": {
                    "type": "string",
                    "example": "100"
                },
                "need": {
                    "type": "string",
                    "example": "钢筋工"
                },
                "project_name": {
                    "type": "string",
                    "example": "主楼建造"
                },
                "required_people": {
                    "type": "string",
                    "example": "11"
                },
                "settlement": {
                    "type": "string",
                    "example": "月薪"
                },
                "type": {
                    "type": "string",
                    "example": "消防"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "model.LocationInfoWrapper": {
            "type": "object",
            "properties": {
                "addr": {
                    "type": "string",
                    "example": "湖北省襄阳市樊城区武商广场店对面人民公园"
                },
                "latitude": {
                    "type": "string",
                    "example": "32.04278"
                },
                "longitude": {
                    "type": "string",
                    "example": "112.15519"
                },
                "title": {
                    "type": "string",
                    "example": "人民广场"
                }
            }
        },
...
    }
}